refactor: move file
Moved
- racingCarGame.js to class folder
