+++ src/index.js
@@ -1,4 +1,4 @@
-import RacingCarGame from "./modules/racingCarGame.js";
+import RacingCarGame from "./class/racingCarGame.js";
 
 const game = new RacingCarGame();
 game.init();
